Cleanup, fix #306
