+++ src/scss/_panktee.scss
@@ -3,6 +3,10 @@
     margin-bottom: -0.25rem;
   }
 
+  .highlight {
+    color: #01669b;
+  }
+
   .transliteration {
     text-transform: capitalize;
   }
